+++ pkg/dbs/studydb/db.go
@@ -78,6 +78,10 @@ func (dbService *StudyDBService) collectionRefConfidentialResponses(instanceID s
 	return dbService.DBClient.Database(dbService.DBNamePrefix + instanceID + "_studyDB").Collection(studyKey + "_confidentialResponses")
 }
 
+func (dbService *StudyDBService) collectionRefResearcherMessages(instanceID string, studyKey string) *mongo.Collection {
+	return dbService.DBClient.Database(dbService.DBNamePrefix + instanceID + "_studyDB").Collection(studyKey + "_researcherMessages")
+}
+
 // DB utils
 func (dbService *StudyDBService) getContext() (ctx context.Context, cancel context.CancelFunc) {
 	return context.WithTimeout(context.Background(), time.Duration(dbService.timeout)*time.Second)