+++ pkg/types/study.go
@@ -15,17 +15,23 @@ const (
 	STUDY_STATUS_INACTIVE = "inactive"
 )
 
+type NotificationSubscription struct {
+	MessageType string `bson:"messageType"`
+	Email       string `bson:"email"`
+}
+
 type Study struct {
-	ID             primitive.ObjectID `bson:"_id,omitempty" json:"id,omitempty"`
-	Key            string             `bson:"key"`
-	SecretKey      string             `bson:"secretKey"`
-	Status         string             `bson:"status"`
-	Members        []StudyMember      `bson:"members"` // users with access to manage study
-	Rules          []Expression       `bson:"rules"`   // defining how the study should run
-	Props          StudyProps         `bson:"props"`
-	NextTimerEvent int64              `bson:"nextTimerEventAfter"`
-	Stats          StudyStats         `bson:"studyStats"`
-	Configs        StudyConfigs       `bson:"configs"`
+	ID                        primitive.ObjectID         `bson:"_id,omitempty" json:"id,omitempty"`
+	Key                       string                     `bson:"key"`
+	SecretKey                 string                     `bson:"secretKey"`
+	Status                    string                     `bson:"status"`
+	Members                   []StudyMember              `bson:"members"` // users with access to manage study
+	Rules                     []Expression               `bson:"rules"`   // defining how the study should run
+	Props                     StudyProps                 `bson:"props"`
+	NextTimerEvent            int64                      `bson:"nextTimerEventAfter"`
+	Stats                     StudyStats                 `bson:"studyStats"`
+	Configs                   StudyConfigs               `bson:"configs"`
+	NotificationSubscriptions []NotificationSubscription `bson:"notificationSubscriptions"`
 }
 
 type StudyMember struct {